requirements.txt
tensorflow
numpy
scikit-learn
google-generativeai
Pillow
python-dotenv